requirements.txt
torch
tqdm
torchsummary
TemporaryDirectory
gdown
numpy
sklearn
PIL
rasterio
matplotlib
seaborn
geopandas
shapely
ee
einops